--- src/consts.ts
@@ -1,6 +1,0 @@
-// Place any global data in this file.
-// You can import this data from anywhere in your site by using the `import` keyword.
-
-export const SITE_TITLE = "Abdul Rafay";
-export const SITE_DESCRIPTION =
-  "Welcome to my universe, Software Engineer and Open Source enthusiast";
