+++ src/pages/404.astro
@@ -1,6 +1,5 @@
 ---
 import BaseHead from "../components/AstroCompoent/base/BaseHead.astro";
-// import { SITE_TITLE, SITE_DESCRIPTION } from "../consts";
 import authorConfig from "../config/info";
 export const prerender = true;
 import SpeedInsights from "@vercel/speed-insights/astro";
