+++ src/pages/about.astro
@@ -1,7 +1,6 @@
 ---
 import BaseHead from "../components/AstroCompoent/base/BaseHead.astro";
 import Header from "../components/AstroCompoent/header/Header.astro";
-// import { SITE_TITLE, SITE_DESCRIPTION } from "../consts";
 import "../styles/home/home-page.css";
 import JobExperience from "../components/AstroCompoent/JobAndExperience/JobExperience.astro";
 import authorConfig from "../config/info";
@@ -25,7 +24,7 @@ import Footer from "../components/AstroCompoent/footer/Footer.astro";
         <div class="left-container">
           <div
             class="lottie-animation-container"
-            style="width: 70%; height: 70%;"
+            style="inline-size: 70%; block-size: 70%;"
           >
             <img src={authorConfig.picture} alt="Profile Image" />
           </div>
