+++ src/pages/blog/index.astro
@@ -2,7 +2,7 @@
 import BaseHead from "../../components/AstroCompoent/base/BaseHead.astro";
 import Header from "../../components/AstroCompoent/header/Header.astro";
 import Footer from "../../components/AstroCompoent/footer/Footer.astro";
-import { SITE_TITLE, SITE_DESCRIPTION } from "../../consts";
+import authorConfig from "../../config/info";
 import { getCollection } from "astro:content";
 import SpeedInsights from "@vercel/speed-insights/astro";
 import BlogSection from "../../components/ReaactCompoents/BlogPost/BlogPost";
@@ -13,7 +13,10 @@ const posts = (await getCollection("blog")).sort(
 
 <html lang="en">
   <head>
-    <BaseHead title={SITE_TITLE} description={SITE_DESCRIPTION} />
+    <BaseHead
+      title={authorConfig.SiteName}
+      description={authorConfig.SiteDescription}
+    />
   </head>
   <body>
     <Header />
@@ -22,152 +25,3 @@ const posts = (await getCollection("blog")).sort(
     <Footer />
   </body>
 </html>
-
-<!-- 
-
- ██████╗ ██╗     ██████╗     ██╗   ██╗███████╗██████╗ ███████╗██╗ ██████╗ ███╗   ██╗
-██╔═══██╗██║     ██╔══██╗    ██║   ██║██╔════╝██╔══██╗██╔════╝██║██╔═══██╗████╗  ██║
-██║   ██║██║     ██║  ██║    ██║   ██║█████╗  ██████╔╝███████╗██║██║   ██║██╔██╗ ██║
-██║   ██║██║     ██║  ██║    ╚██╗ ██╔╝██╔══╝  ██╔══██╗╚════██║██║██║   ██║██║╚██╗██║
-╚██████╔╝███████╗██████╔╝     ╚████╔╝ ███████╗██║  ██║███████║██║╚██████╔╝██║ ╚████║
- ╚═════╝ ╚══════╝╚═════╝       ╚═══╝  ╚══════╝╚═╝  ╚═╝╚══════╝╚═╝ ╚═════╝ ╚═╝  ╚═══╝
-
-██████╗ ██╗   ██╗██████╗ ███████╗     █████╗ ███████╗████████╗██████╗  ██████╗ 
-██╔══██╗██║   ██║██╔══██╗██╔════╝    ██╔══██╗██╔════╝╚══██╔══╝██╔══██╗██╔═══██╗
-██████╔╝██║   ██║██████╔╝█████╗      ███████║███████╗   ██║   ██████╔╝██║   ██║
-██╔═══╝ ██║   ██║██╔══██╗██╔══╝      ██╔══██║╚════██║   ██║   ██╔══██╗██║   ██║
-██║     ╚██████╔╝██║  ██║███████╗    ██║  ██║███████║   ██║   ██║  ██║╚██████╔╝
-╚═╝      ╚═════╝ ╚═╝  ╚═╝╚══════╝    ╚═╝  ╚═╝╚══════╝   ╚═╝   ╚═╝  ╚═╝ ╚═════╝ 
-                                                                               
- ██████╗ ██████╗ ███╗   ███╗██████╗  ██████╗ ███╗   ██╗███████╗███╗   ██╗████████╗        
-██╔════╝██╔═══██╗████╗ ████║██╔══██╗██╔═══██╗████╗  ██║██╔════╝████╗  ██║╚══██╔══╝        
-██║     ██║   ██║██╔████╔██║██████╔╝██║   ██║██╔██╗ ██║█████╗  ██╔██╗ ██║   ██║           
-██║     ██║   ██║██║╚██╔╝██║██╔═══╝ ██║   ██║██║╚██╗██║██╔══╝  ██║╚██╗██║   ██║           
-╚██████╗╚██████╔╝██║ ╚═╝ ██║██║     ╚██████╔╝██║ ╚████║███████╗██║ ╚████║   ██║           
- ╚═════╝ ╚═════╝ ╚═╝     ╚═╝╚═╝      ╚═════╝ ╚═╝  ╚═══╝╚══════╝╚═╝  ╚═══╝   ╚═╝           
-                                                                                          
--->
-<!--
-<!doctype html>
-<html lang="en">
-  <head>
-    <BaseHead title={SITE_TITLE} description={SITE_DESCRIPTION} />
-    <style>
-      main {
-        width: 960px;
-      }
-      ul {
-        display: flex;
-        flex-wrap: wrap;
-        gap: 2rem;
-        list-style-type: none;
-        margin: 0;
-        padding: 0;
-      }
-      ul li {
-        width: calc(50% - 1rem);
-      }
-      ul li * {
-        text-decoration: none;
-        transition: 0.2s ease;
-      }
-      ul li:first-child {
-        width: 100%;
-        margin-bottom: 1rem;
-        text-align: center;
-      }
-      ul li:first-child img {
-        width: 100%;
-      }
-      ul li:first-child .title {
-        font-size: 2.369rem;
-      }
-      ul li img {
-        margin-bottom: 0.5rem;
-        border-radius: 12px;
-      }
-      ul li a {
-        display: block;
-      }
-      .title {
-        margin: 0;
-        color: var(--text-light);
-        line-height: 1;
-      }
-      .date {
-        margin: 0;
-        color: rgb(var(--gray));
-      }
-      ul li a:hover h4,
-      ul li a:hover .date {
-        color: rgb(var(--accent));
-      }
-      ul a:hover img {
-        box-shadow: var(--box-shadow);
-      }
-      @media (max-width: 720px) {
-        ul {
-          gap: 0.5em;
-        }
-        ul li {
-          width: 100%;
-          text-align: center;
-        }
-        ul li:first-child {
-          margin-bottom: 0;
-        }
-        ul li:first-child .title {
-          font-size: 1.563em;
-        }
-        .author {
-          font-size: 1rem;
-
-          margin-top: 0.5rem;
-        }
-      }
-    </style>
-  </head>
-  <body>
-    <Header />
-    <SpeedInsights />
-    <main>
-      <h1>Articles</h1>
-
-      <br />
-      <section>
-        <ul>
-          {
-            posts
-              .filter((post) =>
-                import.meta.env.PROD ? !post.data.draft : true
-              )
-              .sort(
-                (a, b) =>
-                  new Date(b.data.pubDate).valueOf() -
-                  new Date(a.data.pubDate).valueOf()
-              )
-              .map((post) => (
-                <li>
-                  <a href={`/blog/${post.slug}/`}>
-                    <Image
-                      src={post.data.heroImage || ""}
-                      alt="author Avator"
-                      format="avif"
-                      width={720}
-                      height={360}
-                    />
-                    <h4 class="title">{post.data.title}</h4>
-                    <p class="date">
-                      <FormattedDate date={post.data.pubDate} />
-                    </p>
-                    <p class="author date">By {post.data.authorName}</p>
-                  </a>
-                </li>
-              ))
-          }
-        </ul>
-      </section>
-    </main>
-    <Footer />
-  </body>
-</html> -->
