+++ src/pages/newsletter/index.astro
@@ -2,7 +2,7 @@
 import BaseHead from "../../components/AstroCompoent/base/BaseHead.astro";
 import Header from "../../components/AstroCompoent/header/Header.astro";
 import Footer from "../../components/AstroCompoent/footer/Footer.astro";
-import { SITE_TITLE, SITE_DESCRIPTION } from "../../consts";
+import authorConfig from "../../config/info";
 import { getCollection } from "astro:content";
 import SpeedInsights from "@vercel/speed-insights/astro";
 import Newsletter from "../../components/ReaactCompoents/Newsletter/ListNewsletter/NewsletterList";
@@ -12,7 +12,10 @@ const newsletters = await getCollection("newsletter");
 
 <html lang="en">
   <head>
-    <BaseHead title={SITE_TITLE} description={SITE_DESCRIPTION} />
+    <BaseHead
+      title={authorConfig.SiteName}
+      description={authorConfig.SiteDescription}
+    />
   </head>
   <body>
     <Header />
