+++ src/pages/terms-and-conditions.astro
@@ -2,7 +2,6 @@
 import BaseHead from "../components/AstroCompoent/base/BaseHead.astro";
 import Footer from "../components/AstroCompoent/footer/Footer.astro";
 import Header from "../components/AstroCompoent/header/Header.astro";
-import { SITE_TITLE, SITE_DESCRIPTION } from "../consts";
 import SpeedInsights from "@vercel/speed-insights/astro";
 import authorConfig from "../config/info";
 ---
